+++ forum/templates/users/get_captcha.html
@@ -10,11 +10,21 @@
 <main>
     <h1>Test Google reCAPTCHA</h1>
     <form action="/password/reset/" method="POST">
+      {% csrf_token %}
         <label for="email">Email:</label>
         <input type="email" id="email" name="email" placeholder="Enter your email" required>
         <div class="g-recaptcha" data-sitekey='6LcFKJ4qAAAAAJW5zXi4IhWcCm5ZkuFclq9ru3wn'></div>
         <button type="submit">Submit</button>
     </form>
+  <script>
+    document.getElementById('captcha-form').addEventListener('submit', function (event) {
+        const response = grecaptcha.getResponse();
+        if (response.length === 0) {
+            event.preventDefault();
+            document.getElementById('error-message').style.display = 'block';
+        }
+    });
+</script>
 </main>
 </body>
 </html>
